split sims into subclasses
