+++ eee/evolve/base/container.py
@@ -1,14 +1,9 @@
 
-from eee._private.check.eee_variables import check_num_generations
-from eee._private.check.eee_variables import check_mutation_rate
-from eee._private.check.eee_variables import check_population_size
 from eee._private.check.ensemble import check_ensemble
 from eee._private.check.standard import check_int
-from eee._private.interface import run_cleanly
 
 from eee.evolve import FitnessContainer
 from eee.evolve import GenotypeContainer
-from eee.evolve.engine.wright_fisher import wright_fisher
 
 import numpy as np
 
@@ -20,6 +15,9 @@ class SimulationContainer:
     Class for running Wright-Fisher simulation on a thermodynamic ensemble. 
     """
     
+    calc_type = None
+    run = None
+
     def __init__(self,
                  ens,
                  ddg_df,
@@ -106,54 +104,45 @@ class SimulationContainer:
                                      ddg_df=ddg_df,
                                      choice_function=self._rng.choice)
 
-    @property
-    def calc_type(self):
-        return "wright_fisher"
+        # Make sure the SimulationContainer container is subclassed -- on itself
+        if self.__class__ is SimulationContainer:
+            err = "\nOnly subclasses of SimulationContainer should be used\n\n"
+            raise NotImplementedError(err)
 
-    #     # Make sure the SimulationContainer container is subclassed -- on itself
-    #     if self.__class__ is SimulationContainer:
-    #         err = "\nOnly subclasses of SimulationContainer should be used\n\n"
-    #         raise NotImplementedError(err)
+        # Make sure the subclass defines calc_type
+        if self.calc_type is None:
+            err = f"\nsubclasses of {super().__class__} must define the property\n"
+            err += "`calc_type`. This should be a simple string. For example, \n"
+            err += "`cool_sim` would allow eee to recognize 'cool_sim' as\n"
+            err += "referring to this class in a simulation.json file.\n\n"
+            raise NotImplementedError(err)
 
-    #     # Make sure the subclass implements calc_type and run
-    #     self.calc_type
-    #     self.run
+        # Make sure the subclass defines run
+        if self.run is None:
+            err = \
+            """
+            subclasses of SimulationContainer must define the :code:`run`"
+            function. This should be a function that runs the actual
+            simulation. We recommend that the function:
 
-    # @property
-    # def calc_type(self):
-    #     err = f"\nsubclasses of {super().__class__} must define the property\n"
-    #     err += "`calc_type`. This should be a simple string. For example, \n"
-    #     err += "`cool_sim` would allow eee to recognize 'cool_sim' as\n"
-    #     err += "referring to this class in a simulation.json file.\n\n"
-    #     raise NotImplementedError(err)
+            1. Checks the sanity of all input arguments using the 
+            :code:`eee._private.check.standard` functions.
 
-    # @run_cleanly
-    # def run(self,*args,**kwargs):
+            2. Takes an argument :code:`output_directory` and then runs the 
+            calculation in that directory.
+            
+            3. Creates a :code:`calc_params` dictionary that holds the names and
+            values of all run arguments. The function should then pass this
+            dictionary to :code:`self._prepare_calc` before doing its run.
 
-    #     err = f"\nsubclasses of {super().__class__} must define the `run`\n"
-    #     err += "function. This should be a function that runs the actual\n"
-    #     err += "simulation. We recommend that the function:\n\n":
-    #     err += "  1. Checks the sanity of all input arguments using the\n"
-    #     err += "     eee._private.check.standard functions.\n"
-    #     err += "  2. Takes an argument `output_directory` and then runs the\n"
-    #     err += "     calculation in that directory.\n"
-    #     err += "  3. Creates a `calc_params` dictionary that holds the names\n"
-    #     err += "     and values of all input parameters. The function should\n"
-    #     err += "     then pass this dictionary to self._prepare_calc.\n"
-    #     err += "  4. Closes out the calculation after running using\n"
-    #     err += "     self._complete_calc.\n"
-    #     err += "  5. Uses the @run_cleanly decorator, which cleans up the run\n"
-    #     err += "     in the case of a crash.\n\n"
+            4. Closes out the calculation after running using 
+            :code:`self._complete_calc.`
 
-    #     raise NotImplementedError(err)
-    
-    #     # parse arguments and put all of them into "calc_params" dictionary
-    #     self._prepare_calc(output_directory,calc_params)
+            5. Uses the :code:`@run_cleanly` decorator to ensure sane behavior in 
+            the event of a crash. 
+            """
 
-    #     # Actually do calculation here, writing out any important outputs. 
-
-    #     self._compete_calc()
-
+            raise NotImplementedError(err)
 
     def _prepare_calc(self,
                       output_directory,
@@ -240,64 +229,5 @@ class SimulationContainer:
         # Return to starting dir if recorded
         if hasattr(self,"_current_dir"):
             os.chdir(self._current_dir)
-
-    @run_cleanly
-    def run(self,
-            output_directory="eee_sim",
-            population_size=1000,
-            mutation_rate=0.01,
-            num_generations=100,
-            write_prefix="eee_sim",
-            write_frequency=1000):
-        """
-        Run a simulation and save files to an output directory.
-        
-        Parameters
-        ----------
-        output_directory : str, default="eee_sim"
-            do simulation in this output directory
-        population_size : int, default=1000
-            population size for the simulation. Should be > 0.
-        mutation_rate : float, default=0.01
-            mutation rate for the simulation. Should be > 0.
-        num_generations : int, default=100
-            number of generations to run the simulation for
-        write_prefix : str, default="eee_sim"
-            write output files during the run with this prefix. 
-        write_frequency : int, default=1000
-            write the generations out every write_frequency generations. 
-        """
-
-        population_size = check_population_size(population_size)
-        mutation_rate = check_mutation_rate(mutation_rate) 
-        num_generations = check_num_generations(num_generations)
-        write_prefix = f"{write_prefix}"
-
-        write_frequency = check_int(value=write_frequency,
-                                    variable_name="write_frequency",
-                                    minimum_allowed=1)
-    
-        # Record the new keys
-        calc_params = {}
-        calc_params["population_size"] = population_size
-        calc_params["mutation_rate"] = mutation_rate
-        calc_params["num_generations"] = num_generations
-        calc_params["write_prefix"] = write_prefix
-        calc_params["write_frequency"] = write_frequency
-
-        self._prepare_calc(output_directory=output_directory,
-                           calc_params=calc_params)
-        
-        # Run and return a Wright Fisher simulation.
-        self._gc, _ =  wright_fisher(gc=self._gc,
-                                     population=population_size,
-                                     mutation_rate=mutation_rate,
-                                     num_generations=num_generations,
-                                     write_prefix=write_prefix,
-                                     write_frequency=write_frequency,
-                                     rng=self._rng)
-        
-        self._complete_calc()
-
