+++ reports/completeness-crawler.txt
@@ -25,6 +25,8 @@ FOUND tests._private.test_interface.test_launch
 FOUND tests._private.test_interface.test_rmtree
 FOUND tests._private.test_interface.test_run_cleanly
 FOUND tests._private.test_wrap.test_wrap_function
+FOUND tests.evolve.calculation.test_wf_sim.test_WrightFisherSimulation
+FOUND tests.evolve.calculation.test_wf_sim.test_WrightFisherSimulation_run
 FOUND tests.evolve.engine.test_wright_fisher.test__write_outputs
 FOUND tests.evolve.engine.test_wright_fisher.test_wright_fisher
 FOUND tests.evolve.fitness.test_ff.test_ff_neutral
@@ -59,9 +61,9 @@ FOUND tests.evolve.genotype.test_genotype_container.test_GenotypeContainer_mut_e
 FOUND tests.evolve.genotype.test_genotype_container.test_GenotypeContainer_mutate
 FOUND tests.evolve.genotype.test_genotype_container.test_GenotypeContainer_trajectories
 FOUND tests.evolve.genotype.test_genotype_container.test_GenotypeContainer_wt_sequence
-FOUND tests.evolve.simulation.test_load_json.test__validate_calc_kwargs
-FOUND tests.evolve.simulation.test_load_json.test_load_json
 FOUND tests.io.test_load_ddg.test_load_ddg
+FOUND tests.io.test_load_json.test__validate_calc_kwargs
+FOUND tests.io.test_load_json.test_load_json
 FOUND tests.io.test_load_structure.test__load_structure_cif
 FOUND tests.io.test_load_structure.test__load_structure_pdb
 FOUND tests.io.test_load_structure.test_load_structure
@@ -82,7 +84,7 @@ FOUND tests.test_ensemble.test_Ensemble_species
 FOUND tests.test_ensemble.test_Ensemble_to_dict
 FAKE tests._private.test_wrap.test_IterFromFile
 FAKE tests._private.test_wrap.test_IterFromFile___call__
-FAKE tests.evolve.simulation.test_simulate_tree.test_simulate_tree
+FAKE tests.evolve.base.test_simulate_tree.test_simulate_tree
 FAKE tests.evolve.test_analysis.test_extract_alignment
 FAKE tests.evolve.test_analysis.test_get_genotype_frequencies
 FAKE tests.evolve.test_analysis.test_get_most_common
@@ -99,14 +101,13 @@ MISSING tests._private.test_interface.test_WrappedFunctionException
 MISSING tests._private.test_logger.test_log
 MISSING tests.bin.test_eee-simulate-evolution.test_main
 MISSING tests.bin.test_eee-sync-structures.test_main
+MISSING tests.evolve.base.test_container.test_SimulationContainer
+MISSING tests.evolve.base.test_container.test_SimulationContainer__complete_calc
+MISSING tests.evolve.base.test_container.test_SimulationContainer__prepare_calc
+MISSING tests.evolve.base.test_container.test_SimulationContainer__write_calc_params
+MISSING tests.evolve.base.test_simulate_tree.test__simulate_branch
 MISSING tests.evolve.genotype.test_genotype_container.test_GenotypeContainer_to_dict
-MISSING tests.evolve.simulation.test_container.test_SimulationContainer
-MISSING tests.evolve.simulation.test_container.test_SimulationContainer__complete_calc
-MISSING tests.evolve.simulation.test_container.test_SimulationContainer__prepare_calc
-MISSING tests.evolve.simulation.test_container.test_SimulationContainer__write_calc_params
-MISSING tests.evolve.simulation.test_container.test_SimulationContainer_calc_type
-MISSING tests.evolve.simulation.test_container.test_SimulationContainer_run
-MISSING tests.evolve.simulation.test_simulate_tree.test__simulate_branch
+MISSING tests.evolve.test___init__.test__get_allowable_calcs
 MISSING tests.io.test_write_pdb.test_chain
 MISSING tests.io.test_write_pdb.test_last_class
 MISSING tests.test_ensemble.test_Ensemble_get_dG_obs_fast
@@ -117,12 +118,11 @@ MISSING tests.test_epistasis.test_:
 MISSING tests.test_epistasis.test_ep_sign1,
 MISSING tests.test_epistasis.test_of
 FOUND & EXTRA tests._private.check.test_check_ensemble.test_check_ensemble
+FOUND & EXTRA tests.evolve.base.test_simulation_container.test_SimulationContainer
+FOUND & EXTRA tests.evolve.base.test_simulation_container.test_SimulationContainer__complete_calc
+FOUND & EXTRA tests.evolve.base.test_simulation_container.test_SimulationContainer__prepare_calc
+FOUND & EXTRA tests.evolve.base.test_simulation_container.test_SimulationContainer__write_calc_params
 FOUND & EXTRA tests.evolve.genotype.test_genotype_container.test_to_dict
-FOUND & EXTRA tests.evolve.simulation.test_simulation.test_SimulationContainer
-FOUND & EXTRA tests.evolve.simulation.test_simulation.test_SimulationContainer__complete_calc
-FOUND & EXTRA tests.evolve.simulation.test_simulation.test_SimulationContainer__prepare_calc
-FOUND & EXTRA tests.evolve.simulation.test_simulation.test_SimulationContainer__write_calc_params
-FOUND & EXTRA tests.evolve.simulation.test_simulation.test_SimulationContainer_run
 FOUND & EXTRA tests.test_ensemble.test_get_dG_obs_fast
 FOUND & EXTRA tests.test_ensemble.test_get_fx_obs_fast
 FOUND & EXTRA tests.test_ensemble.test_load_mu_dict
