--- tests/evolve/simulation/test_simulate_tree.py
@@ -1,5 +1,0 @@
-
-from eee.evolve.simulation.simulate_tree import simulate_tree
-
-def test_simulate_tree():
-    pass
